The babel config loading is taken to a separate file
Everything regarding loading babel configuration is handled in a single file. So by requiring this file you can get the complete babel config. This should help supporting tools like storybook snapshot tests to get users babel config.
