+++ src/server/config.js
@@ -2,50 +2,10 @@
 
 import fs from 'fs';
 import path from 'path';
-import JSON5 from 'json5';
+import loadBabelConfig from './babel_config';
 
 // avoid ESLint errors
 const logger = console;
-
-function removeReactHmre(presets) {
-  const index = presets.indexOf('react-hmre');
-  if (index > -1) {
-    presets.splice(index, 1);
-  }
-}
-
-// Tries to load a .babelrc and returns the parsed object if successful
-function loadBabelConfig(babelConfigPath) {
-  let config;
-  if (fs.existsSync(babelConfigPath)) {
-    const content = fs.readFileSync(babelConfigPath, 'utf-8');
-    try {
-      config = JSON5.parse(content);
-      config.babelrc = false;
-      logger.info('=> Loading custom .babelrc');
-    } catch (e) {
-      logger.error(`=> Error parsing .babelrc file: ${e.message}`);
-      throw e;
-    }
-  }
-
-  if (!config) return null;
-
-  // Remove react-hmre preset.
-  // It causes issues with react-storybook.
-  // We don't really need it.
-  // Earlier, we fix this by runnign storybook in the production mode.
-  // But, that hide some useful debug messages.
-  if (config.presets) {
-    removeReactHmre(config.presets);
-  }
-
-  if (config.env && config.env.development && config.env.development.presets) {
-    removeReactHmre(config.env.development.presets);
-  }
-
-  return config;
-}
 
 // `baseConfig` is a webpack configuration bundled with storybook.
 // React Storybook will look in the `configDir` directory
@@ -53,26 +13,8 @@ function loadBabelConfig(babelConfigPath) {
 export default function (configType, baseConfig, configDir) {
   const config = baseConfig;
 
-  // Search for a .babelrc in the config directory, then the module root
-  // directory. If found, use that to extend webpack configurations.
-  let babelConfig = loadBabelConfig(path.resolve(configDir, '.babelrc'));
-  let inConfigDir = true;
-
-  if (!babelConfig) {
-    babelConfig = loadBabelConfig('.babelrc');
-    inConfigDir = false;
-  }
-
-  if (babelConfig) {
-    // If the custom config uses babel's `extends` clause, then replace it with
-    // an absolute path. `extends` will not work unless we do this.
-    if (babelConfig.extends) {
-      babelConfig.extends = inConfigDir ?
-        path.resolve(configDir, babelConfig.extends) :
-        path.resolve(babelConfig.extends);
-    }
-    config.module.loaders[0].query = babelConfig;
-  }
+  const babelConfig = loadBabelConfig(configDir);
+  config.module.loaders[0].query = babelConfig;
 
   // Check whether a config.js file exists inside the storybook
   // config directory and throw an error if it's not.
